Moving plugins into osgQt, following Robert’s advice:
One of the items that could complicate things is the qfont plugin as this would normally be placed in the lib/osgPlugins-version directory.  This coupling to a specific OSG version is not helpful for the osgQt build or maintainers.

I think the solution is to simply place the src/osgPlugins/qfont/ReaderQFont.cpp file directly into the osgQt directory and remove the separate plugin completely.  One wouldn't use the qfont ReaderWriter without using qt/osgQt so there is nothing to loose from making this change, what you gain is simplicity and flexibility.