Update osgQOpenGLWidget
"When implementing a new widget, it is almost always useful to reimplement sizeHint() to provide a reasonable default size for the widget and to set the correct size policy with setSizePolicy()."   
https://doc.qt.io/qt-5/qwidget.html#size-hints-and-size-policies  

Default value of QWidget::rect is 0,0,640,480. This SD aspect ratio agrees with OSG when widget is shown by itself. However, once widget is put into a layout it has to return a valid sizeHint in order to preserve this aspect ratio and size.